TST: clear more internal pyepics state
clear_cache detaches from the current context and attaches to a new
one, however this does not reset the 'initial_context' that is a
module level global used by pyepics to allow all python-level threads
to share the same underlying libca context.  Setting
`epics.ca.initial_context = None` causes the new context to be
reported as the initial context.

This allows (master branch) of ophyd to be imported later in the
tests.