FIX: Avoid race condition in pyepics_shim PV init.
The connection callback was being run before the __init__ had a
chance to set self._caproto_pv. There was no way to fix this by
reordering the statements in __init__, but since the callback
itself passes caproto_pv in as an argument, we can just get it
from there rather than relying on access to it in instance state.