FIX: Pass around Subscription not subscriptionid.
This saves a lookup to match the sub_id with the Subscription on the
receive side and, more importantly, it avoids this race condition.

1. Server adds Subscription.
2. Server processes EventCancelRequest.
3. ChannelData sends first EventAddRequest into queue.
4. The sub_id in question has already been removed for the caches. BOOM!

By passing around the Subscription itself, we it's OK if it's already
been removed for the caches. We can process it normally anyway. Then,
the subscription-sending loop will check at the last moment whether the
subscription has been canceled and omit the EventAddResponse from the
batch to be sent.