FIX: following epics-base tcpiiu.cpp, use SO_KEEPALIVE and TCP_NODELAY
Sent data is getting buffered and held up at the TCP stack-level, which
can be flushed immediately using TCP_NODELAY.

While some consider TCP_NODELAY a bad idea, it is what epics-base uses
and appears to be an acceptable trade-off to reduce latency.

It may be worthwhile in the future to selectively enable/disable
TCP_NODELAY based on the current conditions.

Ref: NSLS-II/caproto#385
Ref: https://github.com/epics-base/epics-base/blob/30268e35774775f29738b4391028bb07366fb280/modules/ca/src/client/tcpiiu.cpp#L734