FIX: remember to clear ioid cache
This stops a leak of resources and ensures that we can make more than
2**16 ReadNotify and WriteNotify calls.

There is still a failure mode where we get a WriteNotifyRequest /
ReadNotifyRequest in, something goes wrong, we never get a
WriteNotifyRespons / ReadNotifyResponse.  In that case we will leak a
reference to the Channel object so long as the Circuit object is alive.