DOC: Show search release notes when they are actually released.
This occurred because I neglected to continue to do a "backport" release
and omit them. The only reason for leaving them out in the first place
was because they are not yet exactly the same as the reference
implementation. But they are definite improvement, so I think no harm is
done in releasing them.