FIX: correctly handle re-connecting if the server dies
If the circuit is killed, all of the pending events are set so we will
never hit the logic where it was (the `event.wait` timing out after
the circuit died).  There is a small race condition left were the
event times out and then as we are processing to raise the
TimeoutException the circuit dies, but in that case it has timed out
first so we should in fact give priority to that behavior.