FIX: add a lock to sending over a VirtualCircuit
This fixes a bug when:
 - there is a ReadResponse that `sock_sendall` can not send in on go
 - there is a monitor that will fire during that readout

Because the buffer for the ReadResponse is sent in parts, the bytes
for the AddEvent can get sent while the client still thinks it is
receiving the payload from the ReadResponse.  The tail of the payload
will then be read as the next message from the server.

 a) corrupt the read payload
 b) result is a bogus message

curio and trio share a code path, asyncio has it's own.