API/FIX: do not track search requests on server
The search related API was removed from :class:`ca.Broadcaster`, all
of the search request accounting is handled in the client code.  The
code that is used on the servers can not do this book keeping because
we can not know what other servers are out there and if the
SearchRequest actually got serviced (as that goes back uni-cast).

partially addresses #647