FIX: sync client search ID was always 0
If every IOC uses the same CID of 0, communication channels may get
crossed and cause confusion (IOC says PV exists but it doesn't).

It's very plausible this is the cause of ConnectionRefused errors and
other random failures, as the sync client is used in the test suite to
determine if an IOC has started up or not.